Reformatted, renamed methods, removed boilerplate and unused code.
