+++ src/main/java/com/sda/eventapp/repository/UserRepository.java
@@ -3,9 +3,5 @@ package com.sda.eventapp.repository;
 import com.sda.eventapp.model.User;
 import org.springframework.data.repository.CrudRepository;
 
-import java.util.Optional;
-
 public interface UserRepository extends CrudRepository<User, Long> {
-
-    Optional<User> findById(Long id);
 }
