+++ src/main/java/com/sda/eventapp/service/UserService.java
@@ -8,14 +8,9 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class UserService {
-    private final UserRepository repository;
-
-    public User findById(Long id) {
-        return repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User with id" + id + " not found"));
-    }
+    private final UserRepository userRepository;
 
     public User save(User user) {
-        return repository.save(user);
+        return userRepository.save(user);
     }
 }
