+++ src/main/java/com/sda/eventapp/web/mapper/EventMapper.java
@@ -6,12 +6,15 @@ import com.sda.eventapp.web.dto.EventWithBasicData;
 import java.util.List;
 
 public class EventMapper {
-    public static List<EventWithBasicData> toWebpage(List<Event> events){
-        return events.stream().map(event -> EventWithBasicData.builder()
-                .title(event.getTitle())
-                .description(event.getDescription())
-                .startingDateTime(event.getStartingDateTime())
-                .endingDateTime(event.getEndingDateTime())
-                .build()).toList();
+    public static List<EventWithBasicData> toWebpage(List<Event> events) {
+        return events.stream()
+                .map(event -> EventWithBasicData
+                        .builder()
+                        .title(event.getTitle())
+                        .description(event.getDescription())
+                        .startingDateTime(event.getStartingDateTime())
+                        .endingDateTime(event.getEndingDateTime())
+                        .build())
+                .toList();
     }
 }
