+++ src/main/java/com/sda/eventapp/web/mvc/controller/EventController.java
@@ -1,6 +1,5 @@
-package com.sda.eventapp.web.mvc;
+package com.sda.eventapp.web.mvc.controller;
 
-import com.sda.eventapp.model.Event;
 import com.sda.eventapp.service.EventService;
 import com.sda.eventapp.web.mvc.form.CreateEventForm;
 import com.sda.eventapp.web.mvc.mapper.EventMapper;
@@ -12,24 +11,21 @@ import org.springframework.ui.ModelMap;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 
-
 @Slf4j
 @Controller
-@RequestMapping({"/event"})
 @RequiredArgsConstructor
-
+@RequestMapping({"/event"})
 public class EventController {
     private final EventService eventService;
 
-
     @GetMapping("/create")
-    public String createEvent(ModelMap model) {
+    public String create(ModelMap model) {
         model.addAttribute("event", new CreateEventForm());
         return "create-event";
     }
 
     @PostMapping("/create")
-    public String createEvenyByPost(@ModelAttribute("event") @Valid CreateEventForm form, Errors errors) {
+    public String handleCreate(@ModelAttribute("event") @Valid CreateEventForm form, Errors errors) {
         if (errors.hasErrors()) {
             return "create-event";
         }
@@ -38,19 +34,19 @@ public class EventController {
     }
 
     @GetMapping("/update/{id}")
-    public String updateEvent(ModelMap model, @PathVariable Long id) {
+    public String update(ModelMap model, @PathVariable Long id) {
         model.addAttribute("event", new CreateEventForm());
-        Event foundEvent = eventService.findById(id);
-        model.addAttribute("event", foundEvent);
+        model.addAttribute("event", eventService.findById(id));
         return "update-event";
     }
 
     @PostMapping("/update")
-    public String updateEventByPost(@ModelAttribute("event") @Valid CreateEventForm form, Errors errors) {
+    public String handleUpdate(@ModelAttribute("event") @Valid CreateEventForm form, Errors errors) {
         if (errors.hasErrors()) {
             return "update-event";
         }
-        eventService.updateByModify(form);
+        eventService.update(form);
         return "index";
     }
 }
+
